added jsondb

+++ core/leveldb.go
@@ -1,13 +1,28 @@
 package core
 
 import (
+	"encoding/json"
+	"io/ioutil"
+	"runtime"
+
 	"github.com/syndtr/goleveldb/leveldb"
 )
 
-var dbPath = "/var/lib/trojan-manager"
+// LoginInfo Only for Admin to store its information
+type LoginInfo struct {
+	Username string `json: "username"`
+	Password string `json: "password"`
+}
+
+var dbPath = "/var/lib/vpsman-manager"
+var jsonPath = "./vpsman-manager"
 
 // GetValue 获取leveldb值
 func GetValue(key string) (string, error) {
+	if runtime.GOOS == "windows" {
+		GetValueJSON(key)
+	}
+	// linux or macos
 	db, err := leveldb.OpenFile(dbPath, nil)
 	defer db.Close()
 	if err != nil {
@@ -23,6 +38,11 @@ func GetValue(key string) (string, error) {
 // SetValue 设置leveldb值
 // admin的密码是保存在leveldb中
 func SetValue(key string, value string) error {
+	// windows
+	if runtime.GOOS == "windows" {
+		SetValueJSON(key, value)
+	}
+	// linux and macos
 	db, err := leveldb.OpenFile(dbPath, nil)
 	defer db.Close()
 	if err != nil {
@@ -33,10 +53,50 @@ func SetValue(key string, value string) error {
 
 // DelValue 删除值
 func DelValue(key string) error {
+	if runtime.GOOS == "windows" {
+		if err := SetValueJSON("", ""); err != nil {
+			return err
+		}
+		return nil
+	}
 	db, err := leveldb.OpenFile(dbPath, nil)
 	defer db.Close()
 	if err != nil {
 		return err
 	}
 	return db.Delete([]byte(key), nil)
+}
+
+// GetValueJSON 从json文件读取
+func GetValueJSON (key string) (string, error){
+	// windows
+		loginInfo := LoginInfo{}
+		data, err := ioutil.ReadFile(jsonPath)
+		if err != nil {
+			return "", err
+		}
+		if err := json.Unmarshal(data, &loginInfo); err != nil {
+			return "", err
+		}
+		if loginInfo.Username == key {
+			return loginInfo.Password, nil
+		}
+		return "", nil
+	
+}
+
+// SetValueJSON 设置json文件
+func SetValueJSON(key string, value string) error {
+	loginInfo := LoginInfo{
+		Username: key,
+		Password: value,
+	}
+	data, err := json.MarshalIndent(loginInfo, "", "    ");
+	if err != nil {
+		return err
+	}
+	if err := ioutil.WriteFile(jsonPath, data, 0644); err != nil {
+		return err
+	}
+	return nil
 }
